Downloader: module reorg refactoring.
Cleaner separation into submodules.
"headers_downloader" is organized on the top level to prepare for "bodies_downloader".

Extracted parts:
* headers - models like HeaderSlices that all stages and tests use
* ui - views to display headers
* stages - downloader Stage implementations
* verification - HeaderSliceVerifier, PreverifiedHashesConfig-s

Conventions:
* each module exposes with "pub" only parts that are needed to be public in the super module
* if a module needs to access super::super::X, it is aliased with "use super::super::X",
and becomes accessible with "super::X". This helps if the module is moved.

Removed:
* downloader_impl is obsolete: HeadersDownloader is used directly, UISystem and sentry_status_provider are managed outside
* stage_stream is merged into downloader_stage_loop
