--- src/downloader/headers/mod.rs
@@ -1,41 +1,0 @@
-pub mod downloader;
-pub mod header;
-pub mod header_slice_verifier;
-pub mod header_slice_verifier_mock;
-pub mod header_slices;
-
-mod average_delta_counter;
-mod downloader_forky;
-mod downloader_linear;
-mod downloader_preverified;
-mod downloader_stage_loop;
-mod header_slice_status_watch;
-mod parallel;
-mod preverified_hashes_config;
-mod stage;
-mod stage_stream;
-
-mod fetch_receive_stage;
-mod fetch_request_stage;
-mod fork_mode_stage;
-mod penalize_stage;
-mod refetch_stage;
-mod refill_stage;
-mod retry_stage;
-mod save_stage;
-mod top_block_estimate_stage;
-mod verify_stage_forky_link;
-mod verify_stage_linear;
-mod verify_stage_linear_link;
-mod verify_stage_preverified;
-
-#[cfg(feature = "crossterm")]
-pub mod ui_crossterm;
-#[cfg(feature = "crossterm")]
-pub use ui_crossterm::HeaderSlicesView;
-
-#[cfg(not(feature = "crossterm"))]
-pub mod ui_tracing;
-
-#[cfg(not(feature = "crossterm"))]
-pub use ui_tracing::HeaderSlicesView;
