+++ src/downloader/headers_downloader/docs/state_machine_linear.png
