+++ src/downloader/headers_downloader/downloader_stage_loop.rs
@@ -1,10 +1,19 @@
-use super::{
-    header_slices::HeaderSlices,
-    stage_stream::{make_stage_stream, StageStream},
-};
-use std::sync::Arc;
+use super::{headers::header_slices::HeaderSlices, stages::stage::Stage as DownloaderStage};
+use futures_core::Stream;
+use std::{pin::Pin, sync::Arc};
 use tokio_stream::{StreamExt, StreamMap};
 use tracing::*;
+
+type StageStream<'a> = Pin<Box<dyn Stream<Item = anyhow::Result<()>> + 'a + Send>>;
+
+fn make_stage_stream<'a>(mut stage: impl DownloaderStage + 'a) -> StageStream<'a> {
+    let stream = async_stream::stream! {
+        loop {
+            yield stage.execute().await;
+        }
+    };
+    Box::pin(stream)
+}
 
 // Downloading happens with several stages where
 // each of the stages processes blocks in one status,
@@ -25,7 +34,7 @@ impl<'s> DownloaderStageLoop<'s> {
         }
     }
 
-    pub fn insert<Stage: super::stage::Stage + 's>(&mut self, stage: Stage) {
+    pub fn insert<Stage: DownloaderStage + 's>(&mut self, stage: Stage) {
         let name = String::from(std::any::type_name::<Stage>());
         self.stream.insert(name, make_stage_stream(stage));
     }
