+++ src/downloader/headers_downloader/mod.rs
@@ -1,0 +1,16 @@
+pub mod downloader;
+pub mod verification;
+
+use super::ui;
+
+mod headers;
+mod headers_ui;
+mod stages;
+
+mod downloader_forky;
+mod downloader_linear;
+mod downloader_preverified;
+mod downloader_stage_loop;
+
+#[cfg(test)]
+mod downloader_tests;
