+++ src/downloader/headers_downloader/stages/fetch_receive_stage.rs
@@ -1,4 +1,4 @@
-use super::{
+use super::headers::{
     header::BlockHeader,
     header_slices::{HeaderSlice, HeaderSliceStatus, HeaderSlices},
 };
