+++ src/downloader/headers_downloader/stages/save_stage.rs
@@ -1,4 +1,4 @@
-use super::{
+use super::headers::{
     header::BlockHeader,
     header_slice_status_watch::HeaderSliceStatusWatch,
     header_slices::{HeaderSlice, HeaderSliceStatus, HeaderSlices},
